+++ docs/CHANGELOG.rst
@@ -10,6 +10,11 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Added
+-----
+- ``/api/data/:id/parents`` and ``/api/data/:id/children`` API endpoints for
+  listing parents and children Data objects of the object with given ``id``
+
 Fixed
 -----
 - Make sure that Elasticsearch index exists before executing a search query